+++ HISTORY.md
@@ -47,6 +47,7 @@
 
 ### 2025-12-09
 * widgets: added support for combined icon-text labels in tab buttons
+* symbols: added new "blocks" symbol
 
 ### 2025-12-08
 * objects: support letting sprites draw directly on the stage's background